+++ perf-event/CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add `KProbe` and `UProbe` events.
 - Add `Event::update_attrs_with_data` to allow events to store references to
   owned data within `Builder`'s `perf_event_attr` struct.
+- Add `Tracepoint` event type.
 
 ## [0.6.0] - 2023-05-17
 ### Added